Refactor JSONStruct handling by introducing YSON package
