+++ api/converter/converter_test.go
@@ -17,7 +17,6 @@
 package converter_test
 
 import (
-	"fmt"
 	"math"
 	"testing"
 	gotime "time"
@@ -301,280 +300,5 @@ func TestConverter(t *testing.T) {
 		assert.Equal(t, obj.Get("t").(*crdt.Tree).NodeLen(), doc.Root().GetTree("t").NodeLen())
 		assert.Equal(t, obj.Get("t").(*crdt.Tree).Root().Len(), doc.Root().GetTree("t").Len())
 		assert.Equal(t, obj.Get("t").(*crdt.Tree).ToXML(), doc.Root().GetTree("t").ToXML())
-	})
-}
-
-func TestJSONStructConversion(t *testing.T) {
-
-	t.Run("json struct conversion", func(t *testing.T) {
-		doc := document.New("jsonstruct")
-
-		err := doc.Update(func(root *json.Object, p *presence.Presence) error {
-			// an object and primitive types
-			root.SetNewObject("k1").
-				SetNull("k1.0").
-				SetBool("k1.1", true).
-				SetInteger("k1.2", 2147483647).
-				SetLong("k1.3", 9223372036854775807).
-				SetDouble("1.4", 1.79).
-				SetString("k1.5", "4").
-				SetBytes("k1.6", []byte{65, 66}).
-				SetDate("k1.7", gotime.Now()).
-				Delete("k1.5")
-
-			// an array
-			root.SetNewArray("k2").
-				AddNull().
-				AddBool(true).
-				AddInteger(1).
-				AddLong(2).
-				AddDouble(3.0).
-				AddString("4").
-				AddBytes([]byte{65}).
-				AddDate(gotime.Now()).
-				Delete(4)
-
-			// plain text
-			root.SetNewText("k3").
-				Edit(0, 0, "ㅎ").
-				Edit(0, 1, "하").
-				Edit(0, 1, "한").
-				Edit(0, 1, "하").
-				Edit(1, 1, "느").
-				Edit(1, 2, "늘").
-				Edit(2, 2, "구름").
-				Edit(2, 3, "뭉게구")
-
-			// rich text
-			root.SetNewText("k4").
-				Edit(0, 0, "Hello world", nil).
-				Edit(6, 11, "sky", nil).
-				Style(0, 5, map[string]string{"b": "1"})
-
-			// long counter
-			root.SetNewCounter("k5", crdt.LongCnt, 0).
-				Increase(10)
-
-			// integer counter
-			root.SetNewCounter("k6", crdt.IntegerCnt, 0).
-				Increase(10)
-
-			// tree
-			root.SetNewTree("k7").
-				Edit(0, 0, &json.TreeNode{
-					Type: "p",
-					Children: []json.TreeNode{{
-						Type:  "text",
-						Value: "Hello world",
-					}},
-				}, 0)
-
-			return nil
-		})
-		assert.NoError(t, err)
-
-		jsonStruct, err := converter.ToJSONStruct(doc.RootObject())
-		assert.NoError(t, err)
-
-		newDoc := document.New("jsonstruct")
-		err = newDoc.Update(func(root *json.Object, p *presence.Presence) error {
-			root.SetFromJSONStruct(jsonStruct)
-			return nil
-		})
-		assert.NoError(t, err)
-
-		// verify the conversion
-		assert.Equal(t, doc.Marshal(), newDoc.Marshal())
-		newJSONStruct, err := converter.ToJSONStruct(newDoc.RootObject())
-		assert.NoError(t, err)
-		assert.Equal(t, jsonStruct.ToTestString(), newJSONStruct.ToTestString())
-	})
-
-	t.Run("array with nested types test", func(t *testing.T) {
-		doc := document.New("nested-types")
-
-		err := doc.Update(func(root *json.Object, p *presence.Presence) error {
-			arr := root.SetNewArray("nested")
-
-			// Add nested array
-			nestedArr := arr.AddNewArray()
-			nestedArr.AddString("nested1")
-			nestedArr.AddInteger(42)
-
-			// Add nested object
-			obj := arr.AddNewObject()
-			obj.SetString("key", "value")
-			obj.SetNewCounter("counter", crdt.IntegerCnt, 10)
-
-			text := arr.AddNewText()
-			text.Edit(0, 0, "Hello")
-			text.Edit(5, 5, " World")
-			text.Style(0, 5, map[string]string{"bold": "true"})
-
-			// Add nested tree
-			arr.AddNewTree(&json.TreeNode{
-				Type: "p",
-				Children: []json.TreeNode{
-					{Type: "text", Value: "Tree in array"},
-					{
-						Type: "span",
-						Attributes: map[string]string{
-							"style": "color: red",
-						},
-						Children: []json.TreeNode{
-							{Type: "text", Value: "Styled text"},
-						},
-					},
-				},
-			})
-
-			return nil
-		})
-		assert.NoError(t, err)
-
-		// Convert to JSONStruct
-		jsonStruct, err := converter.ToJSONStruct(doc.RootObject())
-		assert.NoError(t, err)
-
-		// Convert back to document
-		newDoc := document.New("nested-types")
-		err = newDoc.Update(func(root *json.Object, p *presence.Presence) error {
-			root.SetFromJSONStruct(jsonStruct)
-			return nil
-		})
-		assert.NoError(t, err)
-
-		// Verify the conversion
-		assert.Equal(t, doc.Marshal(), newDoc.Marshal())
-		newJSONStruct, err := converter.ToJSONStruct(newDoc.RootObject())
-		assert.NoError(t, err)
-		assert.Equal(t, jsonStruct.ToTestString(), newJSONStruct.ToTestString())
-	})
-
-	t.Run("literal JSONStruct conversion test", func(t *testing.T) {
-		jsonStruct := &converter.JSONObjectStruct{
-			JSONType: api.ValueType_VALUE_TYPE_JSON_OBJECT,
-			Value: map[string]converter.JSONStruct{
-				"nested": &converter.JSONArrayStruct{
-					JSONType: api.ValueType_VALUE_TYPE_JSON_ARRAY,
-					Value: []converter.JSONStruct{
-						&converter.JSONArrayStruct{
-							JSONType: api.ValueType_VALUE_TYPE_JSON_ARRAY,
-							Value: []converter.JSONStruct{
-								&converter.JSONPrimitiveStruct{
-									JSONType:  api.ValueType_VALUE_TYPE_STRING,
-									ValueType: crdt.String,
-									Value:     "nested1",
-								},
-								&converter.JSONPrimitiveStruct{
-									JSONType:  api.ValueType_VALUE_TYPE_INTEGER,
-									ValueType: crdt.Integer,
-									Value:     int32(42),
-								},
-							},
-						},
-						&converter.JSONObjectStruct{
-							JSONType: api.ValueType_VALUE_TYPE_JSON_OBJECT,
-							Value: map[string]converter.JSONStruct{
-								"counter": &converter.JSONCounterStruct{
-									JSONType:  api.ValueType_VALUE_TYPE_INTEGER_CNT,
-									ValueType: crdt.IntegerCnt,
-									Value:     int32(10),
-								},
-								"key": &converter.JSONPrimitiveStruct{
-									JSONType:  api.ValueType_VALUE_TYPE_STRING,
-									ValueType: crdt.String,
-									Value:     "value",
-								},
-							},
-						},
-						&converter.JSONTextStruct{
-							JSONType: api.ValueType_VALUE_TYPE_TEXT,
-							Value:    `[{"attrs":{"bold":"true"},"val":"Hello"},{"val":" World"}]`,
-						},
-						&converter.JSONTreeStruct{
-							JSONType: api.ValueType_VALUE_TYPE_TREE,
-							Value: `{"type":"p","children":[{"type":"text","value":"Tree in array"},` +
-								`{"type":"span","children":[{"type":"text","value":"Styled text"}],"attributes":{"style":"color: red"}}]}`,
-						},
-					},
-				},
-			},
-		}
-
-		doc := document.New("literal-jsonstruct")
-		err := doc.Update(func(root *json.Object, p *presence.Presence) error {
-			root.SetFromJSONStruct(jsonStruct)
-			return nil
-		})
-		assert.NoError(t, err)
-		newJsonStruct, err := converter.ToJSONStruct(doc.RootObject())
-		assert.NoError(t, err)
-		assert.Equal(t, jsonStruct.ToTestString(), newJsonStruct.ToTestString())
-	})
-
-	t.Run("unsupported JSONStruct type", func(t *testing.T) {
-		doc := document.New("unsupported-jsonstruct")
-		err := doc.Update(func(root *json.Object, p *presence.Presence) error {
-			defer func() {
-				r := recover()
-				assert.NotNil(t, r)
-				errStr := fmt.Sprintf("%v", r)
-				assert.Contains(t, errStr, "unsupported primitive type")
-			}()
-
-			root.SetFromJSONStruct(&converter.JSONObjectStruct{
-				JSONType: api.ValueType_VALUE_TYPE_JSON_OBJECT,
-				Value: map[string]converter.JSONStruct{
-					"key": &converter.JSONPrimitiveStruct{
-						JSONType:  api.ValueType_VALUE_TYPE_STRING,
-						ValueType: 2000000,
-						Value:     "value",
-					},
-				},
-			})
-			return nil
-		})
-		assert.NoError(t, err)
-		assert.Equal(t, `{}`, doc.Marshal())
-	})
-
-	t.Run("invalid text JSON", func(t *testing.T) {
-		doc := document.New("invalid-json")
-		err := doc.Update(func(root *json.Object, p *presence.Presence) error {
-			defer func() {
-				r := recover()
-				assert.NotNil(t, r)
-				errStr := fmt.Sprintf("%v", r)
-				assert.Contains(t, errStr, "failed to parse text JSON")
-			}()
-
-			text := root.SetNewText("text")
-			text.EditFromJSONStruct(converter.JSONTextStruct{
-				JSONType: api.ValueType_VALUE_TYPE_TEXT,
-				Value:    "invalid json",
-			})
-			return nil
-		})
-		assert.NoError(t, err)
-		assert.Equal(t, `{"text":[]}`, doc.Marshal())
-	})
-
-	t.Run("invalid tree JSON", func(t *testing.T) {
-		doc := document.New("invalid-tree")
-		err := doc.Update(func(root *json.Object, p *presence.Presence) error {
-			treeNode, err := json.GetTreeRootNodeFromJSONStruct(converter.JSONTreeStruct{
-				JSONType: api.ValueType_VALUE_TYPE_TREE,
-				Value:    "invalid json",
-			})
-			if err != nil {
-				return err
-			}
-			root.SetNewTree("tree", treeNode)
-			return nil
-		})
-		assert.Error(t, err)
-		assert.Contains(t, err.Error(), "failed to parse tree JSON")
-		assert.Equal(t, `{}`, doc.Marshal())
 	})
 }
